Refactor: split builtin_tool_coverage.rs into smaller modules (#117)
* Refactor: split builtin_tool_coverage.rs into smaller modules

Split tests/e2e_traces/builtin_tool_coverage.rs (434 lines) into cohesive
submodules to improve readability and reduce merge-conflict surface area:

- common.rs: Shared test helpers (RigConfig, run_trace_test, macro)
- time.rs: Time tool tests (parse, diff, invalid)
- routine.rs: Routine lifecycle tests
- job.rs: Job management tests
- http.rs: HTTP tool tests

The original file now serves as a registry using #[path] attributes,
consistent with the pattern from PR #23.

Fixes #24


* Fix: apply cargo fmt to split modules

Apply formatting fixes to resolve CI lint violations:
- Reorder imports alphabetically
- Collapse single-use function parameters to single line


* Refactor: simplify macro exposure and use fully-qualified path

Address code review feedback:
- Use :: path for run_routine_started_test in macro to avoid
  requiring callers to import the function
- Remove #[macro_export] and use pub(crate) re-export only for cleaner,
  module-local macro usage
- Update routine.rs to import macro directly from common module


* Refactor: move macro into inline module to avoid lint suppression

Address code review feedback:
- Move routine_started_test! macro and its re-export into an inline
   module within common.rs
- This co-locates the macro definition with its re-export, avoiding the
  need for #[allow(unused_imports)] since they're compiled together
- Update routine.rs to import from common::macros::routine_started_test


* Refactor: move Duration import to module level

Move  from inside skill_install_routine_webhook_sim
function to module-level imports for consistency with other imports.


* Fix: use consistent as_u64() for fired_routines check

Update the second occurrence of fired_routines parsing in
skill_install_routine_webhook_sim to use .as_u64() instead of .as_i64()
for consistency with the first occurrence in routine_system_event_emit.


* Refactor: replace macro with proper test functions checking completion

Replace routine_started_test! macro invocations with explicit test
functions that verify tool_calls_completed() and assert ok == true:

- routine_update_delete: Now checks all tools completed successfully
- routine_history: Now checks all tools completed successfully

This detects failures after start, not just that tools were started.
Also removed the now-unused macro, re-export, and run_routine_started_test
function from common.rs.


* Refactor: error propagation, unconditional auto_approve, rstest consolidation

Address code review feedback:

1. Error handling (common.rs):
   - Add HarnessError enum with Display + Error impl
   - Change run_trace_test and run_trace_test_with_timeout to return
     Result instead of panicking with .expect()
   - Update all test functions to return anyhow::Result and use ?

2. Unconditional auto_approve (common.rs):
   - Replace conditional builder.with_auto_approve_tools(true) with
     unconditional .with_auto_approve_tools(config.auto_approve)

3. Consolidate tests (routine.rs):
   - Replace routine_create_list, routine_update_delete, routine_history
     with single rstest parameterised test routine_tools
   - Use #[rstest] with #[case] for fixture_path, message, expected_tools
   - Leave routine_system_event_emit and skill_install_routine_webhook_sim
     unchanged as requested


* Fix: restore builtin tool harness auto-approval defaults

Preserve the test rig's previous default approval behaviour by giving
RigConfig an explicit default with auto_approve enabled. This keeps
HTTP and job coverage traces from stalling while still letting tests
override approval behaviour when they need to.


---------
